backend/requirements.txt
asgiref==3.9.1
Django==3.1.12
django-cors-headers==3.10.1
djangorestframework==3.12.4
djongo==1.3.7
dnspython==2.7.0
mongoengine==0.29.1
psycopg2-binary==2.9.10
pymongo==3.11.4
pytz==2025.2
sqlparse==0.2.4
typing_extensions==4.15.0